Moving Hours helper methods to where they should be.
